Update
	CustomTags (DFN and JS) now supports negative int values
	Fixed an issue introduced with starting locations that would dump new characters in the middle of the ocean
	Added new uox.ini flag for choosing a random starting location for new players (from list in uox.ini)
		RANDOMSTARTINGLOCATION=0	// 0 to disable (default), 1 to enable

+++ source/items.cpp
@@ -222,7 +222,7 @@ bool ApplyItemSection( CItem *applyTo, ScriptSection *toApply )
 				if( !customTagName.empty() && !customTagStringValue.empty() )
 				{
 					customTag.m_Destroy		= FALSE;
-					customTag.m_IntValue	= customTagStringValue.toInt();
+					customTag.m_IntValue = customTagStringValue.toLong();
 					customTag.m_ObjectType	= TAGMAP_TYPE_INT;
 					customTag.m_StringValue	= "";
 					applyTo->SetTag( customTagName, customTag );
